+++ bbknn/__init__.py
@@ -1,16 +1,69 @@
 import numpy as np
 import sys
-from scanpy import settings
-from scanpy import logging as logg
 from annoy import AnnoyIndex
 from sklearn.neighbors import KDTree
 from sklearn.neighbors import DistanceMetric
-from scanpy.neighbors import compute_connectivities_umap
 from scipy.spatial import cKDTree
+from scipy.sparse import coo_matrix
+from umap.umap_ import fuzzy_simplicial_set
+try:
+	from scanpy import settings
+	from scanpy import logging as logg
+except ImportError:
+	pass
 try:
 	import faiss
 except ImportError:
 	pass
+
+def get_sparse_matrix_from_indices_distances_umap(knn_indices, knn_dists, n_obs, n_neighbors):
+	'''
+	Copied out of scanpy.neighbors
+	'''
+	rows = np.zeros((n_obs * n_neighbors), dtype=np.int64)
+	cols = np.zeros((n_obs * n_neighbors), dtype=np.int64)
+	vals = np.zeros((n_obs * n_neighbors), dtype=np.float64)
+
+	for i in range(knn_indices.shape[0]):
+		for j in range(n_neighbors):
+			if knn_indices[i, j] == -1:
+				continue  # We didn't get the full knn for i
+			if knn_indices[i, j] == i:
+				val = 0.0
+			else:
+				val = knn_dists[i, j]
+
+			rows[i * n_neighbors + j] = i
+			cols[i * n_neighbors + j] = knn_indices[i, j]
+			vals[i * n_neighbors + j] = val
+
+	result = coo_matrix((vals, (rows, cols)),
+									  shape=(n_obs, n_obs))
+	result.eliminate_zeros()
+	return result.tocsr()
+
+def compute_connectivities_umap(knn_indices, knn_dists,
+		n_obs, n_neighbors, set_op_mix_ratio=1.0,
+		local_connectivity=1.0):
+	'''
+	Copied out of scanpy.neighbors
+	
+	This is from umap.fuzzy_simplicial_set [McInnes18]_.
+	Given a set of data X, a neighborhood size, and a measure of distance
+	compute the fuzzy simplicial set (here represented as a fuzzy graph in
+	the form of a sparse matrix) associated to the data. This is done by
+	locally approximating geodesic distance at each point, creating a fuzzy
+	simplicial set for each such point, and then combining all the local
+	fuzzy simplicial sets into a global one via a fuzzy union.
+	'''
+	X = coo_matrix(([], ([], [])), shape=(n_obs, 1))
+	connectivities = fuzzy_simplicial_set(X, n_neighbors, None, None,
+										  knn_indices=knn_indices, knn_dists=knn_dists,
+										  set_op_mix_ratio=set_op_mix_ratio,
+										  local_connectivity=local_connectivity)
+	distances = get_sparse_matrix_from_indices_distances_umap(knn_indices, knn_dists, n_obs, n_neighbors)
+
+	return distances, connectivities.tocsr()
 
 def create_tree(data,approx,metric,use_faiss,n_trees):
 	'''
@@ -148,7 +201,7 @@ def trimming(cnts,trim):
 		cnts = cnts.T.tocsr()
 	return cnts
 
-def bbknn(adata, batch_key='batch', save_knn=False, copy=False, **kwargs):
+def bbknn(adata, batch_key='batch', approx=True, metric='angular', copy=False, **kwargs):
 	'''
 	Batch balanced KNN, altering the KNN procedure to identify each cell's top neighbours in
 	each batch separately instead of the entire cell pool with no accounting for batch.
@@ -167,8 +220,11 @@ def bbknn(adata, batch_key='batch', save_knn=False, copy=False, **kwargs):
 	n_pcs : ``int``, optional (default: 50)
 		How many principal components to use in the analysis.
 	trim : ``int`` or ``None``, optional (default: ``None``)
-		If not ``None``, trim the neighbours of each cell to these many top connectivities.
-		May help with population independence and improve the tidiness of clustering.
+		Trim the neighbours of each cell to these many top connectivities. May help with 
+		population independence and improve the tidiness of clustering. The lower the value the
+		more independent the individual populations, at the cost of more conserved batch effect.
+		If ``None``, sets the parameter value automatically to 10 times the total number of
+		neighbours for each cell. Set to 0 to skip.
 	approx : ``bool``, optional (default: ``True``)
 		If ``True``, use annoy's approximate neighbour finding. This results in a quicker run time 
 		for large datasets while also potentially increasing the degree of batch correction.
@@ -191,17 +247,18 @@ def bbknn(adata, batch_key='batch', save_knn=False, copy=False, **kwargs):
 		>>> neighbors.KDTree.valid_metrics
 		['p', 'chebyshev', 'cityblock', 'minkowski', 'infinity', 'l2', 'euclidean', 'manhattan', 'l1']
 		>>> pass_this_as_metric = neighbors.DistanceMetric.get_metric('minkowski',p=3)
-	bandwidth : ``float``, optional (default: 1)
-		``scanpy.neighbors.compute_connectivities_umap`` parameter, higher values result in a
-		gentler slope of the connectivities exponentials (i.e. larger connectivity values being returned)
+	set_op_mix_ratio : ``float``, optional (default: 1)
+		UMAP connectivity computation parameter, float between 0 and 1, controlling the 
+		blend between a connectivity matrix formed exclusively from mutual nearest neighbour
+		pairs (0) and a union of all observed neighbour relationships with the mutual pairs
+		emphasised (1)
 	local_connectivity : ``int``, optional (default: 1)
-		``scanpy.neighbors.compute_connectivities_umap`` parameter, how many nearest neighbors of
-		each cell are assumed to be fully connected (and given a connectivity value of 1)
-	save_knn : ``bool``, optional (default: ``False``)
-		If ``True``, save the indices of the nearest neighbours for each cell in ``adata.uns['bbknn']``.
+		UMAP connectivity computation parameter, how many nearest neighbors of each cell
+		are assumed to be fully connected (and given a connectivity value of 1)
 	copy : ``bool``, optional (default: ``False``)
 		If ``True``, return a copy instead of writing to the supplied adata.
 	'''
+	logg.info('computing batch balanced neighbors', r=True)
 	adata = adata.copy() if copy else adata
 	#basic sanity checks to begin
 	#is our batch key actually present in the object?
@@ -210,14 +267,20 @@ def bbknn(adata, batch_key='batch', save_knn=False, copy=False, **kwargs):
 	#do we have a computed PCA? (the .dtype.fields is because of how adata.obsm is formatted)
 	if 'X_pca' not in adata.obsm.dtype.fields:
 		raise ValueError("`adata.obsm['X_pca']` doesn't exist. Run `sc.pp.pca` first.")
+	#metric sanity checks
+	if approx and metric not in ['angular', 'euclidean', 'manhattan', 'hamming']:
+		logg.warn('unrecognised metric for type of neighbor calculation, switching to angular')
+		metric = 'angular'
+	elif not approx and not (metric=='euclidean' or isinstance(metric,DistanceMetric) or metric in KDTree.valid_metrics):
+		logg.warn('unrecognised metric for type of neighbor calculation, switching to euclidean')
+		metric = 'euclidean'
 	#prepare bbknn_pca_matrix input
 	pca = adata.obsm['X_pca']
 	batch_list = adata.obs[batch_key].values
 	#call BBKNN proper
-	bbknn_out = bbknn_pca_matrix(pca=pca,batch_list=batch_list,save_knn=save_knn,**kwargs)
-	#optionally save knn_indices
-	if save_knn:
-		adata.uns['bbknn'] = bbknn_out[2]
+	bbknn_out = bbknn_pca_matrix(pca=pca, batch_list=batch_list,
+								 approx=approx, metric=metric, **kwargs)
+	logg.info('	finished', time=True, end=' ' if settings.verbosity > 2 else '\n')
 	adata.uns['neighbors'] = {}
 	#we'll have a zero distance for our cell of origin, and nonzero for every other neighbour computed
 	adata.uns['neighbors']['params'] = {'n_neighbors': len(bbknn_out[0][0,:].data)+1, 'method': 'umap'}
@@ -225,19 +288,18 @@ def bbknn(adata, batch_key='batch', save_knn=False, copy=False, **kwargs):
 	adata.uns['neighbors']['connectivities'] = bbknn_out[1]
 	logg.hint(
 		'added to `.uns[\'neighbors\']`\n'
-		'    \'distances\', weighted adjacency matrix\n'
-		'    \'connectivities\', weighted adjacency matrix')
+		'	\'distances\', weighted adjacency matrix\n'
+		'	\'connectivities\', weighted adjacency matrix')
 	return adata if copy else None
 
 def bbknn_pca_matrix(pca, batch_list, neighbors_within_batch=3, n_pcs=50, trim=None, 
 		  approx=True, n_trees=10, use_faiss=True, metric='angular', 
-		  bandwidth=1, local_connectivity=1, save_knn=False):
+		  set_op_mix_ratio=1, local_connectivity=1):
 	'''
 	Scanpy-independent BBKNN variant that runs on a PCA matrix and list of per-cell batch assignments instead of
 	an AnnData object. Non-data-entry arguments behave the same way as ``bbknn.bbknn()``.
 	Returns a ``(distances, connectivities)`` tuple, like what would have been stored in the AnnData object.
-	The connectivities are the actual neighbourhood graph. If ``save_knn=True``, the tuple also
-	includes the nearest neighbour indices for each cell as a third element.
+	The connectivities are the actual neighbourhood graph.
 	
 	Input
 	-----
@@ -246,19 +308,18 @@ def bbknn_pca_matrix(pca, batch_list, neighbors_within_batch=3, n_pcs=50, trim=N
 	batch_list : ``numpy.array`` or ``list``
 		A list of batch assignments for each cell.
 	'''
-	logg.info('computing batch balanced neighbors', r=True)
 	#more basic sanity checks/processing
 	#do we have the same number of cells in pca and batch_list?
 	if pca.shape[0] != len(batch_list):
 		raise ValueError("Different cell counts indicated by `pca.shape[0]` and `len(batch_list)`.")
 	#convert batch_list to np.array of strings for ease of mask making later
 	batch_list = np.asarray([str(i) for i in batch_list])
-	#metric sanity checks
+	#metric sanity checks (duplicating the ones in bbknn(), but without scanpy logging)
 	if approx and metric not in ['angular', 'euclidean', 'manhattan', 'hamming']:
-		logg.warn('unrecognised metric for type of neighbor calculation, switching to angular')
+		print('unrecognised metric for type of neighbor calculation, switching to angular')
 		metric = 'angular'
 	elif not approx and not (metric=='euclidean' or isinstance(metric,DistanceMetric) or metric in KDTree.valid_metrics):
-		logg.warn('unrecognised metric for type of neighbor calculation, switching to euclidean')
+		print('unrecognised metric for type of neighbor calculation, switching to euclidean')
 		metric = 'euclidean'
 	#obtain the batch balanced KNN graph
 	knn_distances, knn_indices = get_graph(pca=pca,batch_list=batch_list,n_pcs=n_pcs,n_trees=n_trees,
@@ -270,12 +331,33 @@ def bbknn_pca_matrix(pca, batch_list, neighbors_within_batch=3, n_pcs=50, trim=N
 	knn_distances = knn_distances[np.arange(np.shape(knn_distances)[0])[:,np.newaxis],newidx] 
 	#this part of the processing is akin to scanpy.api.neighbors()
 	dist, cnts = compute_connectivities_umap(knn_indices, knn_distances, knn_indices.shape[0], 
-											 knn_indices.shape[1], bandwidth=bandwidth, 
+											 knn_indices.shape[1], set_op_mix_ratio=set_op_mix_ratio,
 											 local_connectivity=local_connectivity)
-	#optional trimming
-	if trim:
+	#trimming. compute default range if absent
+	if trim is None:
+		trim = 10 * knn_distances.shape[1]
+	#skip trimming if set to 0, otherwise trim
+	if trim > 0:
 		cnts = trimming(cnts=cnts,trim=trim)
-	logg.info('    finished', time=True, end=' ' if settings.verbosity > 2 else '\n')
-	if save_knn:
-		return (dist, cnts, knn_indices)
 	return (dist, cnts)
+
+def extract_cell_connectivity(adata, cell, key='extracted_cell_connectivity'):
+	'''
+	Helper post-processing function that extracts a single cell's connectivity and stores
+	it in ``adata.obs``, ready for plotting. Connectivities range from 0 to 1, the higher
+	the connectivity the closer the cells are in the neighbour graph. Cells with a
+	connectivity of 0 are unconnected in the graph.
+	
+	Input
+	-----
+	adata : ``AnnData``
+		After having BBKNN ran on it.
+	cell : ``str``
+		The name of the cell to extract the connectivities for.
+	key : ``str``, optional (default "extracted_cell_connectivity")
+		What name to store the connectivities under in ``adata.obs``.
+	'''
+	if cell not in adata.obs_names:
+		ValueError('The specified cell is not present in the object.')
+	index = np.arange(len(adata.obs_names))[adata.obs_names==cell][0]
+	adata.obs[key] = np.asarray(adata.uns['neighbors']['connectivities'][index,:].todense())[0]